fix(supabase): schema.sql e RODAR_NO_SUPABASE.sql reintroduziam o bug do position
A migration consertou o banco em producao, mas os dois snapshots de schema continuavam declarando position integer. O job db do CI monta o banco de teste a partir de supabase/schema.sql, e nao das migrations, e o RODAR_NO_SUPABASE.sql e o que se roda num Supabase novo — qualquer ambiente montado por um dos dois nasceria com o bug de volta.

+++ supabase/schema.sql
@@ -1724,7 +1724,9 @@ create table if not exists public.leads (
   source text not null default 'manual',
   utm jsonb not null default '{}'::jsonb,
   stage_id text not null default 'new',
-  position integer not null default 0,
+  -- bigint, e nao integer: recebe Date.now() em milissegundos (~1,787 trilhao),
+  -- que estoura o teto do integer e faz o Postgres recusar o insert inteiro.
+  position bigint not null default 0,
   waitlist_id uuid references public.waitlist(id) on delete set null,
   user_id uuid references auth.users(id) on delete set null,
   created_at timestamptz not null default now(),
